base UI elements fully functional and butter smooth. Animators segregated into their own class
