+++ BlocSpot/app/src/main/java/com/example/peter/blocspot/ui/activity/MapsActivity.java
@@ -19,12 +19,10 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewTreeObserver;
-import android.view.animation.Animation;
-import android.view.animation.LinearInterpolator;
-import android.view.animation.Transformation;
 import android.widget.ActionMenuView;
 
 import com.example.peter.blocspot.R;
+import com.example.peter.blocspot.ui.animations.BlocSpotAnimator;
 import com.google.android.gms.location.LocationListener;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -47,7 +45,7 @@ public class MapsActivity extends AppCompatActivity implements OnMapReadyCallbac
     private LatLng user;
     private LatLng targetPOI;
 
-    static int targetHeight = 0;
+    public static int targetHeight = 0;
 
     private boolean mNotifyIsOn, windowIsOpen;
 
@@ -213,11 +211,11 @@ public class MapsActivity extends AppCompatActivity implements OnMapReadyCallbac
         } else {
             if(toggleMenuPressed(item.getTitle().toString()) && item != this.add){
                 if(windowIsOpen) {
-                    collapse(view);
+                    BlocSpotAnimator.collapse(view);
                     Handler handler = new Handler();
                     handler.postDelayed(new Runnable() {
                         public void run() {
-                            expand(view);
+                            BlocSpotAnimator.expand(view);
                         }
                     }, 350);
                     offsetCenterMapOnPoint(user, SLOWER_CAMERA_SPEED);
@@ -226,45 +224,18 @@ public class MapsActivity extends AppCompatActivity implements OnMapReadyCallbac
                 else {
                     System.out.println("HURR DURR");
                     offsetCenterMapOnPoint(user, STANDARD_CAMERA_SPEED);
-                    expand(view);
+                    BlocSpotAnimator.expand(view);
                     windowIsOpen = true;
                 }
             } else {
                 centerMapOnPoint(user, STANDARD_CAMERA_SPEED);
                 activeMenu = "";
                 setIndicatorsToDark();
-                collapse(view);
+                BlocSpotAnimator.collapse(view);
                 windowIsOpen = false;
             }
         }
         return super.onOptionsItemSelected(item);
-    }
-
-    public static void expand(final View v) {
-
-        // Older versions of android (pre API 21) cancel animations for views with a height of 0.
-        v.getLayoutParams().height = 1;
-        v.setVisibility(View.VISIBLE);
-        Animation a = new Animation()
-        {
-            @Override
-            protected void applyTransformation(float interpolatedTime, Transformation t) {
-                int height = (int)(targetHeight * interpolatedTime);
-                v.getLayoutParams().height = height;
-                v.requestLayout();
-            }
-
-            @Override
-            public boolean willChangeBounds() {
-                return true;
-            }
-        };
-
-        // 1dp/ms
-        float density =  v.getContext().getResources().getDisplayMetrics().density;
-        a.setDuration(250);
-        a.setInterpolator(new LinearInterpolator());
-        v.startAnimation(a);
     }
 
     @Override
@@ -275,35 +246,8 @@ public class MapsActivity extends AppCompatActivity implements OnMapReadyCallbac
             centerMapOnPoint(user, STANDARD_CAMERA_SPEED);
             activeMenu = "";
             setIndicatorsToDark();
-            collapse(view);
+            BlocSpotAnimator.collapse(view);
         }
-    }
-
-    public static void collapse(final View v) {
-        final int initialHeight = v.getMeasuredHeight();
-
-        Animation a = new Animation()
-        {
-            @Override
-            protected void applyTransformation(float interpolatedTime, Transformation t) {
-                if(interpolatedTime == 1){
-                    v.setVisibility(View.GONE);
-                }else{
-                    v.getLayoutParams().height = initialHeight - (int)(initialHeight * interpolatedTime);
-                    v.requestLayout();
-                }
-            }
-
-            @Override
-            public boolean willChangeBounds() {
-                return true;
-            }
-        };
-
-        // 1dp/ms
-        a.setDuration(250);
-        //a.setDuration((int)(initialHeight / v.getContext().getResources().getDisplayMetrics().density));
-        v.startAnimation(a);
     }
 
     public void setupEvenlyDistributedToolbar(){
