+++ BlocSpot/app/src/main/java/com/example/peter/blocspot/ui/animations/BlocSpotAnimator.java
@@ -1,0 +1,64 @@
+package com.example.peter.blocspot.ui.animations;
+
+import android.view.View;
+import android.view.animation.Animation;
+import android.view.animation.LinearInterpolator;
+import android.view.animation.Transformation;
+
+import com.example.peter.blocspot.ui.activity.MapsActivity;
+
+public class BlocSpotAnimator {
+    public static void expand(final View v) {
+
+        // Older versions of android (pre API 21) cancel animations for views with a height of 0.
+        v.getLayoutParams().height = 1;
+        v.setVisibility(View.VISIBLE);
+        Animation a = new Animation()
+        {
+            @Override
+            protected void applyTransformation(float interpolatedTime, Transformation t) {
+                int height = (int)(MapsActivity.targetHeight * interpolatedTime);
+                v.getLayoutParams().height = height;
+                v.requestLayout();
+            }
+
+            @Override
+            public boolean willChangeBounds() {
+                return true;
+            }
+        };
+
+        // 1dp/ms
+        float density =  v.getContext().getResources().getDisplayMetrics().density;
+        a.setDuration(250);
+        a.setInterpolator(new LinearInterpolator());
+        v.startAnimation(a);
+    }
+
+    public static void collapse(final View v) {
+        final int initialHeight = v.getMeasuredHeight();
+
+        Animation a = new Animation()
+        {
+            @Override
+            protected void applyTransformation(float interpolatedTime, Transformation t) {
+                if(interpolatedTime == 1){
+                    v.setVisibility(View.GONE);
+                }else{
+                    v.getLayoutParams().height = initialHeight - (int)(initialHeight * interpolatedTime);
+                    v.requestLayout();
+                }
+            }
+
+            @Override
+            public boolean willChangeBounds() {
+                return true;
+            }
+        };
+
+        // 1dp/ms
+        a.setDuration(250);
+        //a.setDuration((int)(initialHeight / v.getContext().getResources().getDisplayMetrics().density));
+        v.startAnimation(a);
+    }
+}
